changed: move ReplayGain cache to ApplicationVolumeHandling
