+++ xbmc/application/ApplicationVolumeHandling.h
@@ -10,6 +10,7 @@
 
 class CAction;
 class CApplicationPlayer;
+class CSettings;
 
 /*!
  * \brief Class handling application support for audio volume management.
@@ -17,6 +18,16 @@ class CApplicationPlayer;
 class CApplicationVolumeHandling
 {
 public:
+  // replay gain settings struct for quick access by the player multiple
+  // times per second (saves doing settings lookup)
+  struct ReplayGainSettings
+  {
+    int iPreAmp;
+    int iNoGainPreAmp;
+    int iType;
+    bool bAvoidClipping;
+  };
+
   explicit CApplicationVolumeHandling(CApplicationPlayer& appPlayer);
 
   float GetVolumePercent() const;
@@ -27,6 +38,8 @@ public:
   void SetMute(bool mute);
   void ToggleMute(void);
 
+  const ReplayGainSettings& GetReplayGainSettings() const { return m_replayGainSettings; }
+
   static constexpr float VOLUME_MINIMUM = 0.0f; // -60dB
   static constexpr float VOLUME_MAXIMUM = 1.0f; // 0dB
   static constexpr float VOLUME_DYNAMIC_RANGE = 90.0f; // 60dB
@@ -34,6 +47,8 @@ public:
 protected:
   bool IsMutedInternal() const { return m_muted; }
   void ShowVolumeBar(const CAction* action = nullptr);
+
+  void CacheReplayGainSettings(const CSettings& settings);
 
   void Mute();
   void UnMute();
@@ -45,4 +60,5 @@ protected:
   CApplicationPlayer& m_appPlayer; //!< Reference to application player
   bool m_muted = false;
   float m_volumeLevel = VOLUME_MAXIMUM;
+  ReplayGainSettings m_replayGainSettings;
 };
